refactor: fixed code
