+++ .eslintrc.json
@@ -21,7 +21,9 @@
         "max-len":"off",
         "no-mixed-operators":"off",
         "no-bitwise":"off",
-        "eqeqeq":"off"
+        "eqeqeq":"off",
+        "no-plusplus":"off",
+        "no-undef":"off"
     }
 }
