+++ UI/main_row.html
@@ -80,35 +80,6 @@
         </section>
         <section class="cards">
             <div class="cards__wrapper wrapper" id="cards__wrapper">
-                <!-- <div class="cards-container todo">
-                    <div class="cards__info">
-                        <div class="cards__count">10</div>
-                        <div class="cards__title">TO DO</div>
-                    </div>
-                    <div class="card__wrapper-to768" id="todo">                 
-                    </div>
-                </div>
-                <div class="cards-container inprogress" >
-                    <div  class="cards__info">
-                        <div class="cards__count">10</div>
-                        <div class="cards__title">IN PROGRES</div>
-                    </div>
-
-                    <div class="card__wrapper-to768" id="inprogress">
-    
-                        
-                    </div>
-                </div>
-                <div class="cards-container complete" >
-                    <div class="cards__info">
-                        <div class="cards__count">10</div>
-                        <div class="cards__title">COMPLETE</div>
-                    </div>
-
-                    <div class="card__wrapper-to768" id="complete">
-                     
-                    </div>
-                </div> -->
                 <button class="button_show-more">Show more task</button>
             </div>
         </section>
