--- src/model/index.js
@@ -1,3 +1,0 @@
-import TaskArray from "./TaskArray";
-import TaskModel from "./TaskModel";
-export { TaskArray, TaskModel };
