--- src/view/index.js
@@ -1,5 +1,0 @@
-import HeaderView from "./HeaderView";
-import FilterView from "./FilterView";
-import TaskFeedView from "./TaskFeedView";
-import TaskView from "./TaskView";
-export { HeaderView, FilterView, TaskFeedView, TaskView };
